requirements.txt
pytest>=7.0
pytest-timeout>=2.1
pylint>=2.12
gpiozero>=1.6.2
gql>=3.0.0
aiohttp >= 3.8.0
grove.py >= 0.6
pyserial >= 3.5
firebase-admin >= 5.2.0
